listArtworksByPrompt moved to the promptController from artworkController)
